+++ lib/utilFunctions.ts
@@ -1,0 +1,7 @@
+export const sanitizeFileName = (name: string): string => {
+  return name
+    .normalize('NFD')
+    .replace(/[\u0300-\u036f]/g, '')
+    .replace(/[^a-zA-Z0-9]/g, '_')
+    .toLowerCase();
+};
